Recategorized some recent nodes
