+++ backend/src/nodes/nodes/image/create_color_rgb.py
@@ -7,14 +7,16 @@ from ...node_factory import NodeFactory
 from ...properties import expression
 from ...properties.inputs import NumberInput, SliderInput
 from ...properties.outputs import ImageOutput
-from . import category as ImageUtilityCategory
+from . import category as ImageCategory
 
 
-@NodeFactory.register("chainner:image:create_color_rgba")
+@NodeFactory.register("chainner:image:create_color_rgb")
 class CreateColorNode(NodeBase):
     def __init__(self):
         super().__init__()
-        self.description = "Create an image of specified dimensions filled with a specified RGBA color."
+        self.description = (
+            "Create an image of specified dimensions filled with a specified RGB color."
+        )
         self.inputs = [
             NumberInput("Width", minimum=1, unit="px", default=1),
             NumberInput("Height", minimum=1, unit="px", default=1),
@@ -39,30 +41,23 @@ class CreateColorNode(NodeBase):
                 default=126,
                 gradient=["#000000", "#0000ff"],
             ),
-            SliderInput(
-                "Alpha",
-                minimum=0,
-                maximum=255,
-                default=126,
-                gradient=["#000000", "#ffffff"],
-            ),
         ]
         self.outputs = [
             ImageOutput(
                 image_type=expression.Image(
                     width="Input0",
                     height="Input1",
-                    channels="4",
+                    channels="3",
                 )
             )
         ]
-        self.category = ImageUtilityCategory
-        self.name = "Create Color (RGBA)"
+        self.category = ImageCategory
+        self.name = "Create Color (RGB)"
         self.icon = "MdFormatColorFill"
-        self.sub = "Create Images"
+        self.sub = "Make Images"
 
     def run(
-        self, width: int, height: int, red: int, green: int, blue: int, alpha: int
+        self, width: int, height: int, red: int, green: int, blue: int
     ) -> np.ndarray:
-        img = np.full((height, width, 4), (blue, green, red, alpha), dtype=np.uint8)
+        img = np.full((height, width, 3), (blue, green, red), dtype=np.uint8)
         return (img / 255.0).astype(np.float32)
